requirements.txt
python-dotenv
Flask==2.3.3
Flask-Cors==4.0.0
Pillow==10.0.0
google-generativeai==0.3.0
deep-translator==1.11.4
gunicorn==21.2.0